Update VAE example
