+++ examples/vae/models.py
@@ -1,0 +1,57 @@
+"""VAE model definitions."""
+
+from flax import linen as nn
+from jax import random
+import jax.numpy as jnp
+
+
+class Encoder(nn.Module):
+  """VAE Encoder."""
+  latents: int
+
+  @nn.compact
+  def __call__(self, x):
+    x = nn.Dense(500, name='fc1')(x)
+    x = nn.relu(x)
+    mean_x = nn.Dense(self.latents, name='fc2_mean')(x)
+    logvar_x = nn.Dense(self.latents, name='fc2_logvar')(x)
+    return mean_x, logvar_x
+
+
+class Decoder(nn.Module):
+  """VAE Decoder."""
+
+  @nn.compact
+  def __call__(self, z):
+    z = nn.Dense(500, name='fc1')(z)
+    z = nn.relu(z)
+    z = nn.Dense(784, name='fc2')(z)
+    return z
+
+
+class VAE(nn.Module):
+  """Full VAE model."""
+  latents: int = 20
+
+  def setup(self):
+    self.encoder = Encoder(self.latents)
+    self.decoder = Decoder()
+
+  def __call__(self, x, z_rng):
+    mean, logvar = self.encoder(x)
+    z = reparameterize(z_rng, mean, logvar)
+    recon_x = self.decoder(z)
+    return recon_x, mean, logvar
+
+  def generate(self, z):
+    return nn.sigmoid(self.decoder(z))
+
+
+def reparameterize(rng, mean, logvar):
+  std = jnp.exp(0.5 * logvar)
+  eps = random.normal(rng, logvar.shape)
+  return mean + eps * std
+
+
+def model(latents):
+  return VAE(latents=latents)
