+++ examples/vae/requirements.txt
@@ -1,7 +1,7 @@
-absl-py==1.0.0
-flax==0.3.6
-numpy==1.22.0
-optax==0.1.0
-Pillow==9.3.0
-tensorflow==2.9.3
-tensorflow-datasets==4.4.0
+absl-py==1.4.0
+flax==0.6.9
+numpy==1.23.5
+optax==0.1.5
+Pillow==9.5.0
+tensorflow==2.12.0
+tensorflow-datasets==4.9.2
